Extract execution query helpers
